[cherry-pick] semantic analyzer from e93bc7b
