+++ src/manda/parser/semantic_analyzer.c
@@ -6,31 +6,18 @@
 /*   By: sehwjang <sehwjang@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/06 14:07:09 by sehwjang          #+#    #+#             */
-/*   Updated: 2024/04/08 17:51:10 by sehwjang         ###   ########.fr       */
+/*   Updated: 2024/04/10 19:42:56 by sehwjang         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 #include "semantic_analyzer.h"
 
-void traversal_words(t_list **cmd_list, t_list **redi_list, t_parse_tree *parse_tree);
-
-
-t_syntax_tree	*semantic_analyzer(t_parse_tree *parse_tree)
-{
-	t_syntax_tree	*syntax_tree;
-	//int				idx;
-
-	//idx = 0;
-	syntax_tree = (t_syntax_tree *)ft_malloc(sizeof(t_syntax_tree));
-	syntax_tree = make_syntax_tree(parse_tree);
-	return (syntax_tree);
-}
-
+//make syntax tree recursively
 t_syntax_tree	*make_syntax_tree(t_parse_tree *parse_tree)
 {
 	t_syntax_tree	*new_node;
-	
+
 	if (parse_tree->child_type[MID] == terminal)
 	{
 		new_node = syntax_tree_new(((t_token *)parse_tree->child[MID])->type);
@@ -38,7 +25,7 @@ t_syntax_tree	*make_syntax_tree(t_parse_tree *parse_tree)
 		new_node->child[R] = make_syntax_tree(parse_tree->child[RIGHT]);
 		return (new_node);
 	}
-	if(is_subshell(parse_tree))
+	if (is_subshell(parse_tree))
 		return (make_subshell_node(parse_tree));
 	else if (parse_tree->type == simple_command)
 	{
@@ -56,37 +43,70 @@ t_syntax_tree	*make_syntax_tree(t_parse_tree *parse_tree)
 	return (NULL);
 }
 
-void	traversal_words(t_list **cmd_list, t_list **redi_list, t_parse_tree *parse_tree)
-{
-	if (parse_tree->type == io_redirect)
-	{
-		make_redi_list(redi_list, parse_tree);
-		return ;
-	}
-	if (parse_tree->child_type[0] == terminal)
-	{
-		ft_lstadd_back(cmd_list, ft_lstnew(((t_token *)parse_tree->child[0])->word));
-		return ;
-	}
-	traversal_words(cmd_list, redi_list, parse_tree->child[0]);
-	if (parse_tree->child_type[1] == terminal)
-	{
-		ft_lstadd_back(cmd_list, ft_lstnew(((t_token *)parse_tree->child[1])->word));
-		return ;
-	}
-	if (parse_tree->child[1] != NULL)
-		traversal_words(cmd_list, redi_list, parse_tree->child[1]);
-}
-
+//make redirection list recursively
 void	make_redi_list(t_list **redi_list, t_parse_tree *parse_tree)
 {
-
 	if (parse_tree->type == io_redirect)
 	{
 		ft_lstadd_back(redi_list, ft_lstnew(make_redi_node(parse_tree)));
 		return ;
 	}
-	make_redi_list(redi_list, parse_tree->child[0]);
-	if (parse_tree->child[1] != NULL)
-		make_redi_list(redi_list, parse_tree->child[1]);
+	make_redi_list(redi_list, parse_tree->child[LEFT]);
+	if (parse_tree->child[MID] != NULL)
+		make_redi_list(redi_list, parse_tree->child[MID]);
+}
+
+//when meets simple_command, returns a node with L : cmd_list, R : redi_list
+void	make_word_list(t_list **cmd, t_list **redi, t_parse_tree *parse_tree)
+{
+	t_list	*new_node;
+
+	if (parse_tree->type == io_redirect)
+	{
+		make_redi_list(redi, parse_tree);
+		return ;
+	}
+	if (parse_tree->child_type[LEFT] == terminal)
+	{
+		new_node = ft_lstnew(((t_token *)parse_tree->child[LEFT])->word);
+		ft_lstadd_back(cmd, new_node);
+		return ;
+	}
+	make_word_list(cmd, redi, parse_tree->child[LEFT]);
+	if (parse_tree->child_type[MID] == terminal)
+	{
+		new_node = ft_lstnew(((t_token *)parse_tree->child[MID])->word);
+		ft_lstadd_back(cmd, new_node);
+		return ;
+	}
+	if (parse_tree->child[MID] != NULL)
+		make_word_list(cmd, redi, parse_tree->child[1]);
+}
+
+//free parse tree recursively
+void	free_parse_tree(t_parse_tree *parse_tree)
+{
+	int	idx;
+
+	idx = 0;
+	while (idx < 3)
+	{
+		if (parse_tree->child_type[idx] == terminal)
+			free(parse_tree->child[idx]);
+		else if (parse_tree->child_type[idx] == non_terminal)
+			free_parse_tree(parse_tree->child[idx]);
+		idx ++;
+	}
+	free(parse_tree);
+	return ;
+}
+
+//semantic_anaylzer_main
+t_syntax_tree	*semantic_analyzer(t_parse_tree *parse_tree)
+{
+	t_syntax_tree	*syntax_tree;
+
+	syntax_tree = make_syntax_tree(parse_tree);
+	free_parse_tree(parse_tree);
+	return (syntax_tree);
 }
