+++ src/manda/parser/semantic_analyzer_utils.c
@@ -3,15 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   semantic_analyzer_utils.c                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sehwanii <sehwanii@student.42.fr>          #+#  +:+       +#+        */
+/*   By: sehwjang <sehwjang@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024-04-09 06:38:39 by sehwanii          #+#    #+#             */
-/*   Updated: 2024-04-09 06:38:39 by sehwanii         ###   ########.fr       */
+/*   Created: 2024/04/09 06:38:39 by sehwanii          #+#    #+#             */
+/*   Updated: 2024/04/10 19:34:00 by sehwjang         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "semantic_analyzer.h"
-#include "stdbool.h"
+
+//make a new node of ast
 t_syntax_tree	*syntax_tree_new(t_termi type)
 {
 	t_syntax_tree	*new_node;
@@ -30,30 +31,35 @@ t_syntax_tree	*syntax_tree_new(t_termi type)
 	return (new_node);
 }
 
+//make cmd node with child[L] : cmd_list, child[R] : redi_list
 void	make_cmd_node(t_syntax_tree *node, t_parse_tree *parse_tree)
 {
 	char	**cmds;
 	t_list	*cmd_list;
+	t_list	*tmp;
 	t_list	*redi_list;
 	int		size;
 
 	cmd_list = NULL;
 	redi_list = NULL;
-	traversal_words(&cmd_list, &redi_list, parse_tree);
+	make_word_list(&cmd_list, &redi_list, parse_tree);
 	size = ft_lstsize(cmd_list);
 	cmds = (char **)ft_malloc(sizeof(char *) * (size + 1));
 	cmds[size] = NULL;
 	size = 0;
+	tmp = cmd_list;
 	while (cmd_list)
 	{
 		cmds[size++] = cmd_list->content;
+		tmp = cmd_list;
 		cmd_list = cmd_list->next;
+		free(tmp);
 	}
-
-	node->child[0] = (void *)cmds;
-	node->child[1] = redi_list;
+	node->child[L] = (void *)cmds;
+	node->child[R] = redi_list;
 }
 
+//makes new redirection node
 void	*make_redi_node(t_parse_tree *parse_tree)
 {
 	t_parse_tree	*io;
@@ -75,6 +81,7 @@ void	*make_redi_node(t_parse_tree *parse_tree)
 	return (redi_node);
 }
 
+//if parse_tree->child[LEFT] == lparen, return 1
 bool	is_subshell(t_parse_tree *parse_tree)
 {
 	t_parse_tree	*child;
@@ -86,11 +93,13 @@ bool	is_subshell(t_parse_tree *parse_tree)
 		return (false);
 }
 
+//makes subshell_node with child[L] = syntax_tree, child[R] = redi_list
 t_syntax_tree	*make_subshell_node(t_parse_tree *parse_tree)
 {
 	t_syntax_tree	*new_node;
 	t_list			*redi_list;
 	t_parse_tree	*sub_shell_tree;
+	t_parse_tree	*redi_tree;
 
 	sub_shell_tree = (t_parse_tree *)parse_tree->child[LEFT];
 	if (parse_tree->child[MID])
@@ -98,13 +107,15 @@ t_syntax_tree	*make_subshell_node(t_parse_tree *parse_tree)
 		redi_list = NULL;
 		new_node = syntax_tree_new(lparen);
 		new_node->child[L] = make_syntax_tree(sub_shell_tree->child[MID]);
-		make_redi_list(&redi_list, ((t_parse_tree *)parse_tree->child[MID])->child[0]);
+		redi_tree = (t_parse_tree *)parse_tree->child[MID];
+		make_redi_list(&redi_list, redi_tree->child[0]);
 		new_node->child[R] = redi_list;
 	}
 	else
 	{
 		new_node = syntax_tree_new(lparen);
 		new_node->child[L] = make_syntax_tree(sub_shell_tree->child[MID]);
+		new_node->child[R] = NULL;
 	}
 	return (new_node);
 }
