+++ src/manda/parser/syntax_lr_stack.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   lr_stack.c                                         :+:      :+:    :+:   */
+/*   syntax_lr_stack.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: taerakim <taerakim@student.42.fr>          +#+  +:+       +#+        */
+/*   By: taerakim <taerakim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/28 16:43:41 by taerakim          #+#    #+#             */
-/*   Updated: 2024/04/01 15:10:42 by taerakim         ###   ########.fr       */
+/*   Updated: 2024/04/10 18:00:34 by taerakim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
