+++ src/manda/parser/syntax_tree.h
@@ -6,7 +6,7 @@
 /*   By: taerakim <taerakim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/30 17:05:51 by taerakim          #+#    #+#             */
-/*   Updated: 2024/04/10 14:28:29 by taerakim         ###   ########.fr       */
+/*   Updated: 2024/04/12 15:45:12 by taerakim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,22 +24,14 @@ typedef enum e_redi_type
 
 typedef struct s_redi
 {
-	t_redi_type	type;
-	char		*file;
-	//mat_redi		*next;
+	t_redi_type		type;
+	char			*file;
 }				t_redi;
 
 /* ********************************* COMMAND ******************************** */
 # define INIT -1
 # define INFILE 0
 # define OUTFILE 1
-
-//typedef struct s_cmd
-//{
-//	char	**cmds;
-//	t_redi	*redi_list;
-//	int		redi[2];//처음 생성시 INIT(-1)값 넣어주세요
-//}			t_cmd;
 
 /* ****************************** SYNTAX TREE  ****************************** */
 # define L 0
@@ -50,15 +42,13 @@ typedef enum e_symbol
 	sym_or,
 	sym_and,
 	sym_pipe,
-	//sym_only_cmd,//본인 타입
-	sym_command,//좌 char **cmds, 우 t_redi *redi_list
-	sym_subshell//좌 and_or, 우 redi_list
+	sym_command,
+	sym_subshell
 }	t_symbol;
 
 typedef struct s_syntax_tree
 {
 	t_symbol	type;
-	//t_symbol	child_type[2];
 	void		*child[2];
 }	t_syntax_tree;
